requirements.txt
fastapi
uvicorn[standard]
pdfplumber
sentence-transformers
faiss-cpu
python-dotenv
requests
transformers
torch
numpy
tqdm
nltk
PyMuPDF
tqdm
accelerate
bitsandbytes
dotenv
faiss-cpu
annoy